1.Contatinated all routes in Main 2. Changed package struture
