--- src/main/scala/com/activegrid/entities/BaseEntity.scala
@@ -1,6 +1,0 @@
-package com.activegrid.entities
-
-/**
-  * Created by sivag on 27/9/16.
-  */
-trait BaseEntity
