--- src/main/scala/com/activegrid/models/SiteServiceImpl.scala
@@ -1,8 +1,0 @@
-package com.activegrid.models
-
-/**
-  * Created by sivag on 18/10/16.
-  */
-class SiteServiceImpl {
-
-}
