+++ src/main/scala/com/imaginea/activegrid/core/models/ExecutionStatus.scala
@@ -1,4 +1,4 @@
-package com.activegrid.models
+package com.imaginea.activegrid.core.models
 
 /**
   * Created by sivag on 20/10/16.
