+++ src/main/scala/com/imaginea/activegrid/core/models/LogConfigUpdater.scala
@@ -1,4 +1,4 @@
-package com.activegrid.models
+package com.imaginea.activegrid.core.models
 
 /**
   * Created by sivag on 7/10/16.
