+++ src/main/scala/com/imaginea/~
@@ -1,0 +1,22 @@
+package com.activegrid
+
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.stream.ActorMaterializer
+
+
+
+
+object Main extends App {
+
+  implicit val system  = ActorSystem();
+  implicit val materializer = ActorMaterializer()
+
+  implicit val executionContext = system.dispatcher;
+
+  val apiHandler = new RouteDefinition();
+
+  val binding = Http().bindAndHandle(handler = apiHandler.catalogRoute, interface = "localhost", port = 5000);
+
+}
